Split out the tf files
main.tf was getting too large so this commit splits it into smaller tf
files around each function area. This makes it a little bit easier to
reason about.

This was was done as precursor to adding windows bastion support which
will add a number of additional variables and resources to this module.